Fix comment

+++ engine/utils/file_logger.h
@@ -51,10 +51,10 @@ class TRANTOR_EXPORT FileLogger : public AsyncFileLogger {
    private:
     FILE* fp_{nullptr};
     uint64_t max_lines_;
-    std::atomic<size_t> totalLines_{0};
+    std::atomic<uint64_t> totalLines_{0};
     std::string file_name_;
     std::deque<std::string> lineBuffer_;
-    std::atomic<size_t> linesWrittenSinceLastTruncate_{0};
+    std::atomic<uint64_t> linesWrittenSinceLastTruncate_{0};
     static const uint64_t TRUNCATE_CHECK_INTERVAL = 1000;
     mutable std::mutex mutex_;
